Fixes Bug 53057: API - permite eliminar unidade de medida default

+++ Sage50c.API.Sample/Sage50c.API.Sample/Controllers/UnitOfMeasureController.cs
@@ -53,6 +53,9 @@ namespace Sage50c.API.Sample.Controllers {
             if (string.IsNullOrEmpty(UnitOfMeasureID)) {
                 throw new Exception("O código da unidade de medição não está preenchido.");
             }
+            if (UnitOfMeasureID.Equals(systemSettings.SystemInfo.ItemDefaultsSettings.ItemDefaultUnit)) {
+                throw new Exception("Não pode eliminar a unidade de medida default.");
+            }
 
             dsoCache.UnitOfMeasureProvider.Delete(UnitOfMeasureID);
             editState = EditState.None;
